build-llvm.py: Allow the user to install a stage 1 only toolchain
Some users do not want to be forced to do multi-stage builds and still
be able to install the compiled toolchain to a specific folder like the
multi-stage builds. This is added as an option as some users (like
myself) only use the stage 1 only flag as a tool for quicking testing
for build/runtimes issues during a bisect or checking to see if a patch
solves a problem (which was the original point of this flag).

As part of this shuffle, the installation print is moved into the
do_multistage_build function.

Fixes: #31
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>