build-llvm.py: Add '--assertions'
There was an issue with https://github.com/llvm/llvm-project/commit/5354c83ece00690b4dbfa47925f8f5a8f33f1d9e that only appears when
assertions are enabled (because it only adds an assertion). Add a way to
turn them on in a release configuration so that compiles are still
relatively quick.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>