build-llvm.py: Convert '--debug' into '--build-type'
Right now, the LLVM tests are only runnable when specifying '--debug'.
However, some users want to be able to run the LLVM tests on a release
build so they don't have to suffer with slower build times. The real
intention behind '--debug' was just the changing of CMAKE_BUILD_TYPE so
convert '--debug' into '--build-type' and unhook the LLVM tests from
that option.

We only save 29 targets (3783 -> 3752) by turning off LLVM_INCLUDE_TESTS
(tested with '--build-stage1-only') so there is no point in adding an
option to enable/disable that option.

LLVM_BUILD_TESTS is not needed to run tests so omit setting a value for
it.

Closes: #42
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>