build-llvm.py: Enable Polly by default
More and more Android kernel developers are starting to enable Polly.
This was recently enabled in AOSP's LLVM builds:

https://android.googlesource.com/toolchain/llvm_android/+/6ba474f9d931f696860cec49b8c1dd62ac1e5a46

Do the same thing in our build. It only adds approximately 150 targets
to the final LLVM stage (comparing --build-stage1-only with and without
this patch, 3714 -> 3870).

While we're at it, alphabetize the projects.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>