github/workflows: Add build workflow
This tests all of the various scripts in this repo:

* build-llvm.py
* build-binutils.py
* clang-version.sh
* kernel/build.sh

To try and avoid false positive failures, we use the latest stable
release branch of LLVM, which is currently release/10.x.

To get it to fit within a GitHub Actions build, we just do a single
stage build, shallow clone the repo, and just target X86. We do slow
down the build a bit with assertions plus check targets but it is
important that those work before we try to actually test building the
kernel. We do not bother with ccache as it is kind of complicated to use
with GitHub Actions and we can do without the caching overhead.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>