build-llvm.py: Discourage '--build-stage1-only' + '--lto='
There is nothing strictly wrong with putting these two options together;
however, ThinLTO requires either lld or ld.gold so it could result in
weird failures when '--build-stage1-only' is used, which fully depends
on the user's host environment.

Add this note so that we can say "don't do that" if they use these
options and file a bug.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>