kernel/build.sh: Account for host tools after LLVM=1
After the switch to LLVM=1, we implicitly assume that clang is being
built with host backend support. Unfortunately, this is not always the
case. Some users may chose to target LLVM only for backends they are
actively using, which does not have to include the host one. For
example,

$ ./build-llvm.py --pgo --targets "ARM;AArch64"

is common in the custom Android kernel scene because there are only ARM
chips. When kernel/build.sh is invoked by build-llvm.py when '--pgo' is
set, we will only build 'ARCH=arm' and 'ARCH=arm64' kernels. When we
were using 'CC=clang', there was no problem because clang was guaranteed
to have support for the target we were building. However, 'LLVM=1' not
only sets $(CC) to clang, it also sets $(HOSTCC) to clang, meaning that
the host tools such as scripts/dtc will be compiled with clang. If there
is no backend support for the host target (such a lack of x86_64 support
in the example above), we will get a rather cryptic message:

error: unable to create target: 'No available targets are compatible with triple "x86_64-unknown-linux-gnu"'
1 error generated.
make[3]: *** [scripts/Makefile.host:103: scripts/basic/fixdep] Error 1

To fix this, see if the host's target is in the list of supported
targets; if it is not, fall back to using GCC for HOSTCC and HOSTCXX.

Fixes: 8fdead8 ("kernel/build.sh: Convert to LLVM=1")
Reported-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>