build-llvm.py: Add '--show-build-commands' flag
There are times where it is helpful to print the commands that are being
run by subprocess.run(), such as figuring out a minimal cmake command to
report an issue upstream. Add this flag so that the build commands can
be seen and copied easily. It is default off to minimize the output of
the script.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>