kernel/build.sh: Update patching logic
Rather than have a singular monolithic patch file for issues, add the
ability to patch based on *.patch files. There is no point in
distinguishing between defconfig/allmodconfig/allyesconfig, a patch is a
patch is a patch. This will make maintaining this much easier as we can
just add and remove .patch files based on the kernel version we support
without any additional machinery.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>