README.md: Clarify that latest version of distribution should be used whenever possible
LLVM has certain minimum version requirements for cmake and the C
compiler (gcc/clang) and the build will error out if the host version
does not satisfy those requirements. This is expected and it is the
user's responsibility to make sure that their environment satisfies
those requirements according to LLVM's documentation.

Add a note that says the latest version of a distribution (in
particular, Debian and Ubuntu) should be used whenever possible so that
there are not unexpected build errors.

Remove the note about ld.lld for Debian and Ubuntu while we are here, it
is available as the "lld" package on Buster and Focal, which are the
latest versions of these distributions respectively.

Closes: https://github.com/ClangBuiltLinux/tc-build/issues/163
Signed-off-by: Nathan Chancellor <nathan@kernel.org>