kernel/build.sh: Properly get GNU as binary based on host
Up until now, the script assumes that it is running on an x86_64 host,
calling 'as' when the prefix is 'x86_64-linux-gnu-'. AArch64 hardware is
getting really powerful so we should not assume that the script is being
run on x86_64 hardware and properly get 'as' based on the host and the
prefix.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>