build-llvm.py: Fix using relative paths for folders
When a relative folder value is passed, it should be appended to the
current working directory, not tc-build's directory. Fix this by always
calling .resolve() as part of every folder's creation, which will
automatically expand relative paths as relative to the current working
directory, which behaves more in line with other utilities.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>