tc-build: Add custom duration printing function
Right now, datetime.timedelta() is used for printing how long something
took. While there is nothing functionally wrong with this, the format it
prints is a little condensed and hard to quickly parse. Having some
separation between the numbers and unit labels can make things clearer.

This changes the format of the current duration like:

  11:58:59

to:

  11h 58m 59s

Signed-off-by: Nathan Chancellor <nathan@kernel.org>