build-llvm.py: Update GOOD_REVISION to 012ea747ed0275c499f69c82ac0f635f4c76f746
This has been qualified on aarch64 and x86_64 hosts using my
llvm-kernel-testing repository and those same hosts successfully booted
a kernel built from this revision on bare metal.

There is a new warning when building arm64 with ThinLTO but the fix from
012ea747ed0275c499f69c82ac0f635f4c76f746 is much more important than
that warning, which is entirely harmless for the moment. There are no
other major regressions that I am aware of.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>