build-llvm.py: Add '--build-targets'
It can be useful to select subtargets, like 'clang' or 'llvm-ar', to
build single tools when issues are isolated to them, which can reduce
build time. This provides tangible benefits when bisecting, as an issue
can be isolated down to a single commit in less time.

As part of this implementation, the build target parameter is added as a
member to LLVM_Builder and changed to a list to be more flexible and fit
in more with how customizing the builder works in the rest of tc_build.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>