tc_build: kernel: Disable all -Werror configurations
With Linux 6.10, the DRM subsystem added a DRM specific -Werror
configuration [1], which can break our builds because subdir-ccflags-y
overrides KCFLAGS, so our KCFLAGS=-Wno-error gets ignored for
drivers/gpu/drm.

Take advantage of KCONFIG_ALLCONFIG and pass it a temporary file that
sets every known Kconfig -Werror option to 'n'. Any new known options
can easily be added in the future. For ease of implementation, this is
only done for the allconfig option, as KCFLAGS=-Wno-error should
continue to work for defconfigs.

Link: https://git.kernel.org/linus/f89632a9e5fa6c4787c14458cd42a9ef42025434 [1]
Signed-off-by: Nathan Chancellor <nathan@kernel.org>