build-llvm.py: Add ability to override "true" arguments
Users (such as myself) may have wrappers around 'build-llvm.py' to set
default arguments but allow additional arguments to be passed. Arguments
that take a value can be overridden by subsequent instances of the
arguments but boolean actions cannot.

Python 3.9 introduced the BooleanOptionalAction action in the argparse
package, which automatically defines '--foo' and '--no-foo', which would
allow one to override '--assertions' with '--no-assertions' for that one
invocation.

For now, start with '--no-assertions', as the presence of the '--no-'
value for the other boolean options is noisy and they are unlikely to
defaulted to with the need to override. This change can be done to
options as use cases appear.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>