tc_build: llvm: Disable linking against libxml2 by default
libxml2 recently changed its ABI and bumped the soname, resulting in
toolchains built with tc-build failing to start because libxml2 is
linked against by default when it is found.

As far as I can tell, libxml2 is only linked against in areas concerned
with Windows such as llvm-mt and the COFF area of lld, so it is not
important for the use cases that tc-build is concerned with. Disable
linking against libxml2 by default to reduce dynamic dependencies; if
folks want it back on, they can request with '--defines'.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>